adding FWF export, * refactoring science europe template
refs #77606, #77607, #79403
